orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pypdf==4.3.1
python-multipart==0.0.9
pyyaml==6.0.2
